service_processor/requirements.txt
google-cloud-pubsub
google-cloud-firestore
